Major changes to storage format and analysis for TXM.
After recent beamtime at Advanced Light Source beamline 5.3.2.1, the
TXM modules need to be extended to include L-edge spectroscopy. This
seems trivial at first but many parts of the TXM code assumed K-edge
behavior and need to be decoupled, least of which is the HDF storage
format. This commit introduces some of these changes, breaking
backwards compatibility with previous experiments, which will be
restored in a future update.